Back-end/requirement.txt
absl-py==1.3.0
anyio==3.6.2
astunparse==1.6.3
bcrypt==4.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
ecdsa==0.18.0
fastapi==0.85.1
flatbuffers==22.10.26
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
jose==1.0.0
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pymongo==4.3.2
pyparsing==3.0.9
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
torch==1.13.0
torch-mtcnn==0.0.7
torchvision==0.14.0
typing_extensions==4.4.0
unicode==2.9
urllib3==1.26.12
uvicorn==0.19.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
